requirements.txt
black==19.3b0
coverage
coveralls
dj-database-url
Django
flake8
gunicorn
django-heroku
django-crispy-forms
django-phonenumber-field
phonenumbers
django-location-field
requests
xmltodict
sodapy
python-craigslist>=1.0.3
markdown
django-filter
pandas
django-leaflet
django-geojson
jsonfield
Pillow
pep8-naming
bs4
attrs
